backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
rembg>=2.0.62
onnxruntime>=1.20.1
opencv-python-headless>=4.10.0.84
numpy>=1.26.4
scikit-learn>=1.6.0
transformers>=4.47.1
torch>=2.5.1
torchvision>=0.20.1
Pillow>=11.0.0
